post mutation auth

+++ react-graphql/back/src/resolvers/post.ts
@@ -1,4 +1,5 @@
-import { intArg, ObjectDefinitionBlock, objectType, stringArg } from "@nexus/schema/dist/core"
+import { intArg, ObjectDefinitionBlock, stringArg } from "@nexus/schema/dist/core"
+import { getUserId } from "../utils"
 
 //Query
 export const post = (t: ObjectDefinitionBlock<"Query">) => t.field('post', {
@@ -14,36 +15,22 @@ export const post = (t: ObjectDefinitionBlock<"Query">) => t.field('post', {
     }
 })
 
-export const longestPostTitle = (t: ObjectDefinitionBlock<"Query">) => t.string('longestPostTitle', {
-    resolve: async (_, args, ctx) => {
-        const list = await ctx.prisma.post.findMany()
-        let title = ''
-        for (const item of list) {
-            if (item.title.length > title.length) {
-                title = item.title
-            }
-        }
-        return title
-    }
-})
-
-
 // Mutation
 export const createPost = (t: ObjectDefinitionBlock<"Mutation">) => t.field('createPost', {
     type: 'Post',
     args: {
         title: stringArg({ nullable: false }),
-        content: stringArg(),
-        authorEmail: stringArg({ nullable: false }),
+        content: stringArg()
     },
-    resolve: (_, { title, content, authorEmail }, ctx) => {
+    resolve: (_, { title, content }, ctx) => {
+        const userId = getUserId(ctx)
         return ctx.prisma.post.create({
             data: {
                 title,
                 content,
-                published: false,
+                published: true,
                 author: {
-                    connect: { email: authorEmail },
+                    connect: { id: Number(userId) },
                 },
             },
         })
@@ -68,11 +55,15 @@ export const deletePost = (t: ObjectDefinitionBlock<"Mutation">) => t.field('del
     type: 'Post',
     nullable: true,
     args: {
-        id: intArg()
+        id: intArg({ required: true })
     },
-    resolve: (_, { id }, ctx) => {
+    resolve: async (_, { id }, ctx) => {
+        const userId = getUserId(ctx)
+        const post = await ctx.prisma.post.findOne({ where: { id } })
+        if (!post) throw new Error('No post')
+        if (post.authorId !== Number(userId)) throw new Error('Id not match')
         return ctx.prisma.post.delete({
-            where: { id: Number(id) }
+            where: { id }
         })
     }
 })
